move integrations shared components to AC ownership
