+++ apps/web/src/app/admin-console/organizations/integrations/integrations.component.ts
@@ -9,9 +9,11 @@ import { Organization } from "@bitwarden/common/admin-console/models/domain/orga
 import { IntegrationType } from "@bitwarden/common/enums";
 
 import { HeaderModule } from "../../../layouts/header/header.module";
-import { FilterIntegrationsPipe, IntegrationGridComponent, Integration } from "../../../shared/";
 import { SharedModule } from "../../../shared/shared.module";
 import { SharedOrganizationModule } from "../shared";
+import { IntegrationGridComponent } from "../shared/components/integrations/integration-grid/integration-grid.component";
+import { FilterIntegrationsPipe } from "../shared/components/integrations/integrations.pipe";
+import { Integration } from "../shared/components/integrations/models";
 
 @Component({
   selector: "ac-integrations",
