+++ apps/web/src/app/admin-console/organizations/shared/components/integrations/index.ts
@@ -1,0 +1,4 @@
+export * from "./integrations.pipe";
+export * from "./integration-card/integration-card.component";
+export * from "./integration-grid/integration-grid.component";
+export * from "./models";
