+++ apps/web/src/app/admin-console/organizations/shared/components/integrations/integration-card/integration-card.component.html
