+++ apps/web/src/app/admin-console/organizations/shared/components/integrations/integration-card/integration-card.component.ts
@@ -15,7 +15,7 @@ import { SYSTEM_THEME_OBSERVABLE } from "@bitwarden/angular/services/injection-t
 import { ThemeType } from "@bitwarden/common/platform/enums";
 import { ThemeStateService } from "@bitwarden/common/platform/theming/theme-state.service";
 
-import { SharedModule } from "../../../shared.module";
+import { SharedModule } from "../../../../../../shared";
 
 @Component({
   selector: "app-integration-card",
