--- apps/web/src/app/shared/components/index.ts
@@ -1,4 +1,0 @@
-export * from "./integrations/integration-card/integration-card.component";
-export * from "./integrations/integration-grid/integration-grid.component";
-export * from "./integrations/integrations.pipe";
-export * from "./integrations/models";
