+++ apps/web/src/app/shared/index.ts
@@ -1,3 +1,2 @@
 export * from "./shared.module";
 export * from "./loose-components.module";
-export * from "./components/index";
