+++ bitwarden_license/bit-web/src/app/secrets-manager/integrations/integrations.component.ts
@@ -1,7 +1,7 @@
 import { Component } from "@angular/core";
 
 import { IntegrationType } from "@bitwarden/common/enums";
-import { Integration } from "@bitwarden/web-vault/app/shared";
+import { Integration } from "@bitwarden/web-vault/app/admin-console/organizations/shared/components/integrations/models";
 
 @Component({
   selector: "sm-integrations",
