+++ bitwarden_license/bit-web/src/app/secrets-manager/integrations/integrations.module.ts
@@ -1,9 +1,7 @@
 import { NgModule } from "@angular/core";
 
-import {
-  IntegrationCardComponent,
-  IntegrationGridComponent,
-} from "@bitwarden/web-vault/app/shared";
+import { IntegrationCardComponent } from "@bitwarden/web-vault/app/admin-console/organizations/shared/components/integrations/integration-card/integration-card.component";
+import { IntegrationGridComponent } from "@bitwarden/web-vault/app/admin-console/organizations/shared/components/integrations/integration-grid/integration-grid.component";
 
 import { SecretsManagerSharedModule } from "../shared/sm-shared.module";
